+++ src/pcap_tool/__init__.py
@@ -5,9 +5,9 @@ from .parser import (
     iter_parsed_frames,
     validate_pcap_file,
 )
-from .models import PcapRecord, ParsedHandle
-from .pdf_report import generate_pdf_report
-from .summary import generate_summary_df, export_summary_excel
+from .core.models import PcapRecord, ParsedHandle
+from .reporting.pdf_report import generate_pdf_report
+from .reporting.summary import generate_summary_df, export_summary_excel
 from .utils import export_to_csv, anonymize_ip
 from .ai import prepare_ai_data
 from .metrics.stats_collector import StatsCollector
