+++ src/pcap_tool/analysis/__init__.py
@@ -1,0 +1,5 @@
+from .performance import PerformanceAnalyzer
+from .errors import ErrorSummarizer
+from .security import SecurityAuditor
+
+__all__ = ["PerformanceAnalyzer", "ErrorSummarizer", "SecurityAuditor"]
