+++ src/pcap_tool/analysis/errors/__init__.py
@@ -1,0 +1,3 @@
+from .error_summarizer import ErrorSummarizer
+
+__all__ = ["ErrorSummarizer"]
