+++ src/pcap_tool/analysis/performance/__init__.py
@@ -1,0 +1,3 @@
+from .performance_analyzer import PerformanceAnalyzer
+
+__all__ = ["PerformanceAnalyzer"]
