+++ src/pcap_tool/analysis/security/__init__.py
@@ -1,0 +1,3 @@
+from .security_auditor import SecurityAuditor
+
+__all__ = ["SecurityAuditor"]
