+++ src/pcap_tool/processors/icmp_processor.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import ipaddress
 from typing import Dict, Any, Optional, TYPE_CHECKING
 
-from ..models import PcapRecord
+from ..core.models import PcapRecord
 from ..parsers.utils import _safe_int
 from . import PacketProcessor
